Auth controller added
